scenario CLI cleanup
