--- cmd/scenariosfmt/.gitignore
@@ -1,2 +1,0 @@
-# binary
-scenariosfmt
