--- cmd/test/.gitignore
@@ -1,2 +1,0 @@
-# binary
-test
