--- cmd/test/scenariosTestLegacy.go
@@ -1,8 +1,0 @@
-package main
-
-import cli "github.com/multiversx/mx-chain-vm-go/cmd/scenariostestcli"
-
-// Legacy executor, still used by the older versions of erdpy.
-func main() {
-	cli.ScenariosCLI()
-}
